Added a ComicVine details link to the overview page [#161]
 * Link opens to a new tab.
 * Changed the ComicVineId to be a String.

+++ comixed-scrapers/src/main/java/org/comixedproject/scrapers/comicvine/actions/ComicVineGetIssueAction.java
@@ -92,7 +92,7 @@ public class ComicVineGetIssueAction extends AbstractComicVineScrapingAction<Lis
         final ComicVineIssue issue = response.getIssues().get(index);
         final ScrapingIssue entry = new ScrapingIssue();
         entry.setIssueNumber(issue.getIssueNumber());
-        entry.setId(issue.getId().intValue());
+        entry.setId(issue.getId());
         entry.setVolumeId(issue.getVolume().getId());
         entry.setName(issue.getVolume().getName());
         entry.setVolumeName(issue.getVolume().getName());
